refactor: promote APP packages
Preparing to split to true micro-services.
https://cloud.google.com/appengine/docs/standard/go/configuration-files
