+++ .travis.yml
@@ -7,7 +7,7 @@ env:
 install:
 - go mod download
 script:
-- go test -race -coverprofile=coverage.txt -covermode=atomic ./app/...
+- go test -race -coverprofile=coverage.txt -covermode=atomic ./...
 after_success:
 - bash <(curl -s https://codecov.io/bash)
 deploy:
