--- app/models/models.go
@@ -1,4 +1,0 @@
-package models
-
-// POST table to store posts in
-const POST = "Post"
