+++ models/models.go
@@ -6,6 +6,20 @@ import (
 	"google.golang.org/appengine/datastore"
 )
 
+// POST table to store posts in
+const POST = "Post"
+
+// Img the model for a post in an image
+type Img struct {
+	URL      string `datastore:"url"      json:"url"   xml:"url,attr"`
+	Title    string `datastore:"title"    json:"title" xml:"title"`
+	Rating   string `datastore:"rating"   json:"-"     xml:"rating"`
+	Category string `datastore:"-"        json:"-"     xml:"category"`
+
+	// TODO: Implement these!
+	IsValid bool `datastore:"is_valid" json:"-"     xml:"-"`
+}
+
 // Post data object
 type Post struct {
 	// Attributes used for generating unique leu
